fix(WWLib): make uint32/sint32 genuinely 32-bit, fixing a Draft-19-flagged hazard
WWLib/bittype.h's uint32/sint32 were `unsigned long`/`signed long` -
64-bit on LP64 (64-bit Linux/macOS), unlike the DWORD/ULONG hazard fixed
in Milestone 2 Step 7 (which needed a _WIN32 platform split to stay
type-identical with real <windows.h>'s own DWORD), uint32/sint32 have no
real Windows counterpart to preserve compatibility with, so this is a
straight, unconditional fix to `unsigned int`/`signed int` - correct
and sufficient on every platform.

Flagged as urgent in Draft 19 (fable review of Milestone 2,
docs/native-port-plan.md): w3d_file.h alone uses uint32 in 123
on-disk-format struct fields - left 64-bit on Linux/macOS, every one of
those would compute a wrong struct size/offset there.

Fixing it surfaced one real, previously-masked bug: WWSaveLoad/
persistfactory.h's SimplePersistFactoryClass::Save did `(uint32)obj`,
narrowing a pointer directly - "lossless" only by accident while uint32
was 64-bit on Linux, a hard -fpermissive compile error once uint32 is
genuinely 32-bit. Fixed with the correct two-step cast through
uintptr_t (obj is never dereferenced from this value - it's a
runtime-only identity token SaveLoadSystemClass::Register_Pointer/
Find_Pointer remap back to a real pointer within the same save/load
session, so truncating to 32 bits is safe by design).

Verified: real MSVC win32 full rebuild (both changed files are
Windows-compiled too - persistfactory.h's fix is a no-op in value on
Windows since pointers are 32-bit there already, but needed real
verification anyway) and WSL2 linux-x64 (only the pre-existing
imagehlp.h/atlbase.h baseline errors, nothing related). Both
RenderTexturedTriangleTest (all 6 checks) and RenderDeviceInitTest
re-run and still pass.

+++ Core/Libraries/Source/WWVegas/WWLib/bittype.h
@@ -41,12 +41,22 @@
 
 typedef unsigned char	uint8;
 typedef unsigned short	uint16;
-typedef unsigned long	uint32;
+// `unsigned int`, not `unsigned long`: same LP64-vs-LLP64 hazard as
+// DWORD/ULONG below (see that comment) - `unsigned long` is 64-bit on
+// 64-bit Linux/macOS, silently breaking every "32-bit" promise this
+// type's name makes. Unlike DWORD/ULONG there is no real <windows.h>
+// counterpart to stay type-identical with here (uint32/sint32 are
+// WWLib-only names), so no platform split is needed - `unsigned int`/
+// `signed int` are correct and sufficient on every platform. w3d_file.h
+// alone uses `uint32` in 123 on-disk-format struct fields (Draft 19,
+// docs/native-port-plan.md) - left un-32-bit here, every one of those
+// would compute a wrong struct size/offset on 64-bit Linux/macOS.
+typedef unsigned int	uint32;
 typedef unsigned int    uint;
 
 typedef signed char		sint8;
 typedef signed short		sint16;
-typedef signed long		sint32;
+typedef signed int	sint32;
 typedef signed int      sint;
 
 typedef float				float32;
